requirements.txt
Pillow==10.2.0
todoist-api-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
rgbmatrix5x5